requirements.txt
# Core Dependencies
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-dotenv>=0.19.0
python-multipart>=0.0.5
python-decouple>=3.4
pydantic>=2.0.0

# Vector Database
pinecone>=3.0.0

# Text Processing
sentence-transformers>=2.2.2
unstructured>=0.10.0
nltk>=3.8.1
tiktoken>=0.4.0
numpy>=1.21.0
pandas>=1.3.0  # For data processing

# Reranking
cohere>=5.5.0  # For reranker API support

# Frontend
jinja2>=3.0.0
aiofiles>=23.2.0

# HTTP Client
httpx>=0.24.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=3.0.0
pydantic-settings>=2.0.0

# Code Quality
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
pylint>=2.17.0

# Utilities
tqdm>=4.65.0  # Progress bars
